rustdf::sim

Module utility

Functions

reshape_prosit_array

helper function to reshape the flat prosit predicted intensity array into a 3D array where: 1st dimension: 29 rows for every potential ion since prosit allows precursor sequences up to 30 amino acids 2nd dimension: 2 columns for B and Y ions 3rd dimension: 3 channels for charge 1, 2, and 3

sequence_to_all_ions

helper function to convert a peptide ion to all possible ions and serialize the result to a json string

sequence_to_all_ions_par