rustdf::sim::utility

Function reshape_prosit_array

pub fn reshape_prosit_array(array: Vec<f64>) -> Vec<Vec<Vec<f64>>>

helper function to reshape the flat prosit predicted intensity array into a 3D array where: 1st dimension: 29 rows for every potential ion since prosit allows precursor sequences up to 30 amino acids 2nd dimension: 2 columns for B and Y ions 3rd dimension: 3 channels for charge 1, 2, and 3

Arguments

array - A vector of f64 representing the flat prosit array

Returns

A 3D vector of f64 representing the reshaped prosit array